Adds recipe generator feature
Implements a recipe generator feature that suggests random recipes based on user-defined criteria.

The generator allows users to specify the number of days for which to generate recipes and the category of recipes to include.  It also avoids suggesting the same recipe multiple times in a generated menu.

A new "Generator" page has been created in the frontend to accommodate this functionality, and a new "/random" endpoint was added to the backend to retrieve random recipes.